// src/core/utils/enums.ts
export enum codeErrorApi {
  INVALID_SP_TOKEN = 'INVALID_SP_TOKEN',
  USER_NOT_A_NEGOTIATOR = 'USER_NOT_A_NEGOTIATOR',
  NEGOTIATOR_NOT_ALLOWED = 'NEGOTIATOR_NOT_ALLOWED',
  EDB_FUSION = 'EDB_FUSION',
  HIGH_ERROR = 'HIGH_ERROR',
  ONE_SUPPLIER_EXISTS = 'ONE_SUPPLIER_EXISTS',
}

export enum codeErrorPurchase {
  CONTRACT_REQUIRED = 'CONTRACT_REQUIRED',
  SHOW_WARNING = 'SHOW_WARNING',
  REQUIRE_APPROVAL = 'REQUIRE_APPROVAL',
  INVALID_TOTAL = 'INVALID_TOTAL',
}

export enum ROLE {
  NEGOCIATOR = 'NEGOCIADOR',
  ANALYST = 'ANALISTA',
  LEGAL = 'LEGAL',
  BUYER = 'COMPRADOR',
}